requirements.txt
bertopic==0.17.4
demoji==1.1.0
duckdb==1.4.3
google_play_scraper==1.2.7
matplotlib==3.10.8
nltk==3.9.2
numpy==2.3.5
pandas==2.3.3
scikit_learn==1.8.0
seaborn==0.13.2
statsmodels==0.14.5
umap_learn==0.5.9.post2
